Thth2

+++ src/main/java/com/semi/tk/free/FreeDelC.java
@@ -14,12 +14,15 @@ import com.semi.jw.Model;
 public class FreeDelC extends HttpServlet {
 	protected void doGet(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
-		
-		Model.loginCheck(request);
-		FreeDAO.deleteFree(request);
-		FreeDAO.getAllFree(request);
-		FreeDAO.paging(1, request);
-		request.setAttribute("contentPage", "jsp/tk/free/free.jsp");
+
+		if (Model.loginCheck(request)) {
+			FreeDAO.deleteFree(request);
+			FreeDAO.getAllFree(request);
+			FreeDAO.paging(1, request);
+			request.setAttribute("contentPage", "jsp/tk/free/free.jsp");
+		} else {
+			request.setAttribute("contentPage", "로그인하세요");
+		}
 		request.getRequestDispatcher("index.jsp").forward(request, response);
 	}
